// src/types/models/replacementSector.ts
export enum DieCutBlockReplacementSector {
  DIECUTBLOCK_SECTOR = "DIECUTBLOCK_SECTOR",
  RUBBERIZING_SECTOR = "RUBBERIZING_SECTOR",
  CNC_SECTOR = "CNC_SECTOR",
  LAMINATION_SECTOR = "LAMINATION_SECTOR",
  PROJECT_SECTOR = "PROJECT_SECTOR",
}

export enum ClicheCorrugatedReplacementSector {
  PREASSEMBLY_SECTOR = "PREASSEMBLY_SECTOR",
  RECORDING_SECTOR = "RECORDING_SECTOR",
  PREPRESS_SECTOR = "PREPRESS_SECTOR",
  CONFERENCE_SECTOR = "CONFERENCE",
}